+++ crates/iceberg/src/inspect/snapshots.rs
@@ -139,8 +139,8 @@ mod tests {
     use expect_test::expect;
     use futures::TryStreamExt;
 
-    use crate::scan::test_utils::TableTestFixture;
     use crate::test_utils::check_record_batches;
+    use crate::test_utils::scan::TableTestFixture;
 
     #[tokio::test]
     async fn test_snapshots_table() {
